feat: putting correct numeric types in the constants.

+++ src/SS.Game/Constants/GUI/Elements/SSliceImageConstants.cs
@@ -2,6 +2,6 @@
 {
     public static class SSliceImageConstants
     {
-        public const int SPRITE_SLICE_SIZE = 32;
+        public const byte SPRITE_SLICE_SIZE = 32;
     }
 }
